requirements.txt
Flask
coqui-stt
ffmpeg-python
